+++ LoanActionsHistory.js
@@ -7,9 +7,16 @@ import KeyValue from '@folio/stripes-components/lib/KeyValue';
 import MultiColumnList from '@folio/stripes-components/lib/MultiColumnList';
 import Pane from '@folio/stripes-components/lib/Pane';
 import Paneset from '@folio/stripes-components/lib/Paneset';
+import Button from '@folio/stripes-components/lib/Button';
+import Callout from '@folio/stripes-components/lib/Callout';
+import IconButton from '@folio/stripes-components/lib/IconButton';
+import PaneMenu from '@folio/stripes-components/lib/PaneMenu';
+import TabButton from '@folio/stripes-components/lib/TabButton';
+
 import { formatDateTime, getFullName } from './util';
 import loanActionMap from './data/loanActionMap';
 import LoanActionsHistoryProxy from './LoanActionsHistoryProxy';
+import withRenew from './withRenew';
 
 /**
  * Detail view of a user's loan.
@@ -32,8 +39,10 @@ class LoanActionsHistory extends React.Component {
     }).isRequired,
     loan: PropTypes.object,
     user: PropTypes.object,
+    patronGroup: PropTypes.object,
     onCancel: PropTypes.func.isRequired,
     onClickUser: PropTypes.func.isRequired,
+    renew: PropTypes.func,
   };
 
   static manifest = Object.freeze({
@@ -50,7 +59,7 @@ class LoanActionsHistory extends React.Component {
       records: 'loans',
       resourceShouldRefresh: true,
       GET: {
-        path: 'loan-storage/loan-history?query=(id=!{loan.id})',
+        path: 'loan-storage/loan-history?query=(id==!{loan.id})',
       },
     },
   });
@@ -58,6 +67,8 @@ class LoanActionsHistory extends React.Component {
   constructor(props) {
     super(props);
     this.connectedProxy = props.stripes.connect(LoanActionsHistoryProxy);
+    this.renew = this.renew.bind(this);
+
     this.state = {
       loanActionCount: 0,
     };
@@ -71,7 +82,7 @@ class LoanActionsHistory extends React.Component {
       loanActions.records[0].id !== loan.id) return;
     if (!userIds.query || userIds.loan.id !== loan.id) {
       const query = loanActions.records
-        .map(r => `id=${r.metaData.updatedByUserId}`).join(' or ');
+        .map(r => `id==${r.metaData.updatedByUserId}`).join(' or ');
       this.props.mutator.userIds.replace({ query, loan });
     }
 
@@ -92,8 +103,22 @@ class LoanActionsHistory extends React.Component {
     this.props.mutator.loanActionsWithUser.replace({ loan, records });
   }
 
+  renew() {
+    this.props.renew(this.props.loan).then(() => this.showCallout());
+  }
+
+  showCallout() {
+    const message = (
+      <span>
+        The loan for <strong>{this.props.loan.item.title}</strong> was successfully <strong>renewed</strong>.
+      </span>
+    );
+
+    this.callout.sendCallout({ message });
+  }
+
   render() {
-    const { onCancel, loan, user, resources: { loanActionsWithUser } } = this.props;
+    const { onCancel, loan, user, patronGroup, resources: { loanActionsWithUser } } = this.props;
     const loanActionsFormatter = {
       Action: la => loanActionMap[la.action],
       'Action Date': la => formatDateTime(la.loanDate),
@@ -102,9 +127,39 @@ class LoanActionsHistory extends React.Component {
       Operator: la => getFullName(la.user),
     };
 
+    const paneHeader = (
+      <Row style={{ width: '100%' }}>
+        <Col xs={1}>
+          <PaneMenu>
+            <IconButton
+              icon="closeX"
+              onClick={onCancel}
+              title="Close Loan Details"
+              ariaLabel="Close Loan Details"
+            />
+          </PaneMenu>
+        </Col>
+        <Col xs={2}>
+          <PaneMenu>
+            <TabButton title="Loans" aria-label="Loans">Loan Details</TabButton>
+          </PaneMenu>
+        </Col>
+        <Col xs={9}>
+          <TabButton title="Loans" aria-label="User Name and Patron Group">
+            {`Borrower: ${getFullName(user)} (${_.upperFirst(patronGroup.group)})`}
+          </TabButton>
+        </Col>
+      </Row>
+    );
+
     return (
       <Paneset isRoot>
-        <Pane id="pane-loandetails" defaultWidth="100%" dismissible onClose={onCancel} paneTitle="Loan Details">
+        <Pane id="pane-loandetails" defaultWidth="100%" dismissible onClose={onCancel} header={paneHeader}>
+          <Row>
+            <Col>
+              <Button buttonStyle="primary" onClick={this.renew}>Renew</Button>
+            </Col>
+          </Row>
           <Row>
             <Col xs={4} >
               <KeyValue label="Title" value={_.get(loan, ['item', 'title'], '')} />
@@ -174,10 +229,11 @@ class LoanActionsHistory extends React.Component {
               contentData={loanActionsWithUser.records}
             />
           }
+          <Callout ref={(ref) => { this.callout = ref; }} />
         </Pane>
       </Paneset>
     );
   }
 }
 
-export default LoanActionsHistory;
+export default withRenew(LoanActionsHistory);