+++ ViewUser.js
@@ -2,7 +2,6 @@ import { cloneDeep, get, omit, differenceBy } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import queryString from 'query-string';
-import transitionToParams from '@folio/stripes-components/util/transitionToParams';
 import Pane from '@folio/stripes-components/lib/Pane';
 import PaneMenu from '@folio/stripes-components/lib/PaneMenu';
 import { Row, Col } from '@folio/stripes-components/lib/LayoutGrid';
@@ -12,7 +11,6 @@ import IfPermission from '@folio/stripes-components/lib/IfPermission';
 import IfInterface from '@folio/stripes-components/lib/IfInterface';
 import { ExpandAllButton } from '@folio/stripes-components/lib/Accordion';
 import IconButton from '@folio/stripes-components/lib/IconButton';
-import removeQueryParam from '@folio/stripes-components/util/removeQueryParam';
 
 import UserForm from './UserForm';
 import LoansHistory from './LoansHistory';
@@ -62,6 +60,7 @@ class ViewUser extends React.Component {
         POST: PropTypes.func.isRequired,
         DELETE: PropTypes.func.isRequired,
       }),
+      query: PropTypes.object.isRequired,
     }),
     match: PropTypes.shape({
       path: PropTypes.string.isRequired,
@@ -89,6 +88,7 @@ class ViewUser extends React.Component {
   };
 
   static manifest = Object.freeze({
+    query: {},
     selUser: {
       type: 'okapi',
       path: 'users/:{id}',
@@ -121,7 +121,7 @@ class ViewUser extends React.Component {
     settings: {
       type: 'okapi',
       records: 'configs',
-      path: 'configurations/entries?query=(module=USERS and configName=profile_pictures)',
+      path: 'configurations/entries?query=(module==USERS and configName==profile_pictures)',
     },
   });
 
@@ -152,8 +152,6 @@ class ViewUser extends React.Component {
     this.onClickViewLoanActionsHistory = this.onClickViewLoanActionsHistory.bind(this);
     this.onClickCloseLoanActionsHistory = this.onClickCloseLoanActionsHistory.bind(this);
     this.onAddressesUpdate = this.onAddressesUpdate.bind(this);
-    this.transitionToParams = transitionToParams.bind(this);
-    this.removeQueryParam = removeQueryParam.bind(this);
     this.handleSectionToggle = this.handleSectionToggle.bind(this);
     this.handleExpandAll = this.handleExpandAll.bind(this);
   }
@@ -164,8 +162,7 @@ class ViewUser extends React.Component {
 
   onClickViewOpenLoans(e) {
     if (e) e.preventDefault();
-    this.transitionToParams({ layer: 'open-loans' });
-
+    this.props.mutator.query.update({ layer: 'open-loans' });
     this.setState({
       viewOpenLoansMode: true,
     });
@@ -173,7 +170,7 @@ class ViewUser extends React.Component {
 
   onClickViewClosedLoans(e) {
     if (e) e.preventDefault();
-    this.transitionToParams({ layer: 'closed-loans' });
+    this.props.mutator.query.update({ layer: 'closed-loans' });
     this.setState({
       viewOpenLoansMode: false,
     });
@@ -181,13 +178,12 @@ class ViewUser extends React.Component {
 
   onClickCloseLoansHistory(e) {
     if (e) e.preventDefault();
-    this.removeQueryParam('layer');
+    this.props.mutator.query.update({ layer: null });
   }
 
   onClickViewLoanActionsHistory(e, selectedLoan) {
     if (e) e.preventDefault();
-    this.transitionToParams({ layer: 'loan', loan: selectedLoan.id });
-
+    this.props.mutator.query.update({ layer: 'loan', loan: selectedLoan.id });
     this.setState({
       selectedLoan,
     });
@@ -196,7 +192,7 @@ class ViewUser extends React.Component {
   onClickCloseLoanActionsHistory(e) {
     if (e) e.preventDefault();
     const layer = this.state.viewOpenLoansMode ? 'open-loans' : 'closed-loans';
-    this.transitionToParams({ layer, loan: null });
+    this.props.mutator.query.update({ layer, loan: null });
     this.setState({
       selectedLoan: {},
     });
@@ -341,10 +337,10 @@ class ViewUser extends React.Component {
     const userFormData = this.getUserFormData(user, addresses, sponsors, proxies, permissions);
 
     return (
-      <Pane id="pane-userdetails" defaultWidth={this.props.paneWidth} paneTitle={getFullName(user)} lastMenu={detailMenu} dismissible onClose={this.props.onClose}>
+      <Pane id="pane-userdetails" defaultWidth={this.props.paneWidth} paneTitle={getFullName(user)} lastMenu={detailMenu} dismissible onClose={this.props.onClose} appIcon={{ app: 'users' }}>
         <Row end="xs"><Col xs><ExpandAllButton accordionStatus={this.state.sections} onToggle={this.handleExpandAll} /></Col></Row>
 
-        <this.connectedUserInfo user={user} patronGroup={patronGroup} settings={settings} stripes={stripes} />
+        <this.connectedUserInfo accordionId="userInformationSection" user={user} patronGroup={patronGroup} settings={settings} stripes={stripes} expanded={this.state.sections.userInformationSection} onToggle={this.handleSectionToggle} />
         <ExtendedInfo accordionId="extendedInfoSection" stripes={stripes} user={user} expanded={this.state.sections.extendedInfoSection} onToggle={this.handleSectionToggle} />
         <ContactInfo accordionId="contactInfoSection" stripes={stripes} user={user} addresses={addresses} addressTypes={this.addressTypes} expanded={this.state.sections.contactInfoSection} onToggle={this.handleSectionToggle} />
         <IfPermission perm="proxiesfor.collection.get">
@@ -414,6 +410,7 @@ class ViewUser extends React.Component {
             user={user}
             loan={this.state.selectedLoan}
             loanid={this.state.selectedLoan.id}
+            patronGroup={patronGroup}
             stripes={stripes}
             onCancel={this.onClickCloseLoanActionsHistory}
             // when navigating away to another user, clear all loan-related state